Fix missing data freshness check
This managed to get left out of PR #48 for reasons that are quite beyond me.
Properly check for the refresh_check_interval at the list level and at the
global level.

Change the hard coded, fallback default to 10 minutes.

Add a test to actually make sure the config setting works